Feat/player   sistemes de vida i presa de dany (#153)
* feat: Player - sistemes de vida i presa de dany

- Sistema de Vida: Implementado el sistema de puntos de vida del jugador (HP), gestión de daño y periodo de invulnerabilidad (i-frames).
- HUD de Vida Dinámico: Creadas las spritesheets y lógica de animación para mostrar la vida en tiempo real con efectos de transición.
- Menú Game Over: Implementado el flujo de muerte y re-aparición. Incluye assets fragmentados del personaje, efectos de partículas visuales (fragmentos) y sistema de 'Continue' integrado con el SaveSystem.
- Ajustes Estéticos de UI: Corregida la posición de 'Music' y 'SFX' en los paneles de opciones para evitar solapamiento, y rediseño de botones de Game Over.
- Estabilidad: Resueltos errores de compilación (x/y undeclared), eliminados warnings de narrowing conversion y corregido crash al reiniciar partida.

* feat: Player - sistemes de vida i presa de dany

- Sistema de Vida: Implementado el sistema de puntos de vida del jugador (HP), gestión de daño y periodo de invulnerabilidad (i-frames).
- HUD de Vida Dinámico: Creadas las spritesheets y lógica de animación para mostrar la vida en tiempo real con efectos de transición.
- Menú Game Over: Implementado el flujo de muerte y re-aparición. Incluye assets fragmentados del personaje, efectos de partículas visuales (fragmentos) y sistema de 'Continue' integrado con el SaveSystem.
- Ajustes Estéticos de UI: Corregida la posición de 'Music' y 'SFX' en los paneles de opciones para evitar solapamiento, y rediseño de botones de Game Over.
- Estabilidad: Resueltos errores de compilación (x/y undeclared), eliminados warnings de narrowing conversion y corregido crash al reiniciar partida.

* Fix merge conflicts and deduplicate logic in Player and Scene

---------

+++ src/Animation.cpp
@@ -27,8 +27,8 @@ void Animation::Update(float dt) {
 
     timeInFrameMs_ += static_cast<int>(dt);
 
-    while (timeInFrameMs_ >= frames_[currentIndex_].durationMs) {
-        timeInFrameMs_ -= frames_[currentIndex_].durationMs;
+    while (timeInFrameMs_ >= frames_[static_cast<size_t>(currentIndex_)].durationMs) {
+        timeInFrameMs_ -= frames_[static_cast<size_t>(currentIndex_)].durationMs;
 
         if (currentIndex_ + 1 < static_cast<int>(frames_.size())) {
             ++currentIndex_;
@@ -48,7 +48,7 @@ void Animation::Update(float dt) {
 
 const SDL_Rect& Animation::GetCurrentFrame() const {
     if (frames_.empty()) return kEmpty_;
-    return frames_[currentIndex_].rect;
+    return frames_[static_cast<size_t>(currentIndex_)].rect;
 }
 
 int Animation::GetFrameCount() const { return static_cast<int>(frames_.size()); }
